+++ R/policy.R
@@ -2,21 +2,18 @@
 #' 
 #' Extracts the policy from a solved POMDP/MDP.
 #' 
-#' A list (one entry per epoch) with the optimal policy. The policy is a data
-#' frame consisting of three parts.
+#' A list (one entry per epoch) with the optimal policy. 
+#' For converged, infinite-horizon problems solutions, a list with only the converged solution is produced.
+#' The policy is a data
+#' frame consisting o:
 #' 
 #' Part 1: The value function with one column per state. (For MDPs this is just
 #' one column with the state).
 #' 
 #' Part 2: One column with the optimal action.
 #' 
-#' Part 3: One column per observation with the index of the row representing
-#' the policy node in the next epoch.
-#' 
 #' @param x A solved POMDP object.
-#' @return A list with the policy for each epoch. %% ~Describe the value
-#' returned %% If it is a LIST, use %% \item{comp1 }{Description of 'comp1'} %%
-#' \item{comp2 }{Description of 'comp2'} %% ...
+#' @return A list with the policy for each epoch. 
 #' @author Michael Hahsler
 #' @keywords graphs
 #' @examples
@@ -42,7 +39,7 @@ policy <- function(x) {
   
   for(i in 1:length(pg)) {  
     pg[[i]] <- cbind(alpha[[i]],
-      pg[[i]][,-1, drop = FALSE])
+      pg[[i]][,"action", drop = FALSE])
   }
   
   pg