+++ tests/testthat/test-solve_POMDP.R
@@ -37,7 +37,7 @@ expect_equal(pg_horizon$action, pg_stepwise$action) # transitions do not work
 context("solve_POMDP and model files")
 
 sol <- solve_POMDP("http://www.pomdp.org/examples/1d.POMDP")
-plot_policy_graph(sol)
+plot_policy_graph(sol, belief = FALSE)
 
 sol <- solve_POMDP("http://www.pomdp.org/examples/cheese.95.POMDP")
 plot_policy_graph(sol, belief = FALSE)
@@ -47,5 +47,5 @@ sol <- solve_POMDP("http://www.pomdp.org/examples/shuttle.95.POMDP",
 plot_policy_graph(sol, belief = FALSE)
 
 sol <- solve_POMDP("http://www.pomdp.org/examples/stand-tiger.95.POMDP")
-plot_policy_graph(sol)
+plot_policy_graph(sol, belief = FALSE)